feature: Base CSV Checking and RVM Addition (#8)
* Initial Adjustments / Not completed

Testing out having RVM built into doorstop / seeing how it would be for automation. Not completed yet / first commits. This is non-functional at the moment.

* Update markdown.py

Commenting out RVM

* Update item.py

Adjusting new RTM values.

* Updating for rtm attributes

* Removing test reqs

Will allow repo to be in pearl_requirements / shorten time (could be a submodule).

* Changing loging level

Trying to just have logging write to a file for ones that are warnings.

* Adding csvcimple

* Update importer.py

* Update importer.py

* Adding no include_graphics option

Edits to latex.py to allow for latex template to not have include_graphics without erroring out.

* Fix to graphic check in template_data

For latex.py, fixing wrapper statement and check for if 'include_graphics' key is present in template_data.

* Removing default doctitle

Having doctitle be able to be set in the template for latex publishing.

* Adding matrix template setting ability for latex

Will check template_data for if the traceability

* Removing tutorial test files

Prevents the need of doorstop needing to be re-cloned and remove for each run.

* Update latex.py

Spelling correction.

* Adding traceability_matrix to default template

Adding the option to the default template for including the traceability_matrix to the published latex documentation.

* Template Adjustments

* Update .gitignore

* Applied edit per #8

Changed not False to is True.

* Removing test file defaults

* Revert "Applied edit per #8"

This reverts commit eac485a5e39edff4fa8be6ba7f1454872153d425.

* u/sfrinaldi/#1-rvm-edits (#15)

* Edits to RVM items in item.py

Correcting comments in loading/setting methods as well.

* Add default sep value to document.py

* Change reorder to false in tree.py

* Revert "Add default sep value to document.py"

This reverts commit 1cb26f176b04431734ac69b4d4f4bcbdc50b3d57.

* Revert "Change reorder to false in tree.py"

This reverts commit f251b936f568aee7205dcbdf081ffc5145ba6e72.

* Add template_data[rvm] to latex.py

* Add rvm to doorstop.yml template

* Add rvm to wrapper in latex.py

* Add booktabs package to latex template

* Add \landscape for RVM section to latex.py

* Add lscape package to latex templates

* Add rotating/adjustbox packages to latex templates

* Edit wrapper for RVM in latex.py

* Testing sidewaystable in latex.py wrapper

* Test to markdown for latex

* Edit RVM wrapper in latex.py

* Update latex templates

* Edits to packages in latex templates

* Remove test markdown setup in doorstop.cls

* Fix to rvm wrapper in latex.py

* Add \sloppy to doorstop.yml

* Package update to doorstop.cls

* Fix to rvm wrapper in latex.py